+++ src/finder/utils.hpp
@@ -2,7 +2,7 @@
 #include <cstdint>
 
 namespace InsertionFinder::Details {
-    constexpr bool bitcount_less_than_2(std::uint32_t n) {
+    constexpr bool bitcount_less_than_2(std::uint64_t n) {
         return (n & (n - 1)) == 0;
     }
 };
